RF: migrate ShapeStim2 to ShapeStim
- try to maintain full backwards compatibility, old ShapeStim becomes BaseShapeStim
- Aperture, Polygon, and Rect use BaseShapeStim
- dynamic individual vertices are supported; can be slow to assign if there are many vertices, due to re-tesselation
